refactor: modernize deps and tooling
- Replace excel4node with exceljs (actively maintained, fixes uuid vuln)
- Add npm override for uuid >= 11.1.1 (0 audit vulnerabilities)
- Replace link-module-alias with native Node.js package.json#imports
- Replace dot-object with inline deep-pick in config.get
- Replace mkdirp with fs.mkdirSync recursive (built-in)
- Remove unused lodash dependency
- Migrate all ~src/ aliases to #src/ across codebase
- Modernize cleanData.js to fs.promises + async/await
- Remove debug console.log from axiosApi 401 retry handler
- Update userAgent to Chrome 136
- Update CI to Node 22 LTS
