+++ package.json
@@ -1,16 +1,20 @@
 {
   "name": "travian-elephant-finder",
   "version": "0.1.0",
+  "imports": {
+    "#src/*": "./src/*",
+    "#root/*": "./*"
+  },
   "dependencies": {
     "axios": "^1.7.9",
     "cheerio": "^1.2.0",
     "cli-progress": "^3.12.0",
-    "dot-object": "^2.1.5",
     "dotenv": "^17.0.0",
-    "excel4node": "^1.8.2",
-    "jsonfile": "^6.1.0",
-    "lodash": "^4.18.1",
-    "mkdirp": "^3.0.1"
+    "exceljs": "^4.4.0",
+    "jsonfile": "^6.1.0"
+  },
+  "overrides": {
+    "uuid": "^11.1.1"
   },
   "engines": {
     "node": ">=18"
@@ -23,18 +27,12 @@
     "collect": "node src/scripts/collectOasysPosition.js",
     "find": "node src/scripts/findAnimal.js",
     "lint": "eslint --fix --ext .js src/",
-    "test": "echo \"No test specified\"",
-    "postinstall": "link-module-alias"
+    "test": "echo \"No test specified\""
   },
   "devDependencies": {
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.32.0",
-    "link-module-alias": "^1.2.0"
-  },
-  "_moduleAliases": {
-    "~root": ".",
-    "~src": "src/"
+    "eslint-plugin-import": "^2.32.0"
   }
 }
