+++ src/services/util.js
@@ -1,14 +1,13 @@
 const fs = require('fs');
 const path = require('path');
-const mkdirp = require('mkdirp');
-const config = require('~src/config');
+const config = require('#src/config');
 
 const randomIntFromInterval = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
 
 const distance = (x1, y1, x2, y2) => Math.sqrt(((x1 - x2) ** 2) + ((y1 - y2) ** 2));
 
 const createFile = (filename) => {
-  mkdirp.sync(path.dirname(filename));
+  fs.mkdirSync(path.dirname(filename), { recursive: true });
   if (!fs.existsSync(filename)) {
     fs.writeFileSync(filename, '');
   }
